+++ test/test-runner.ts
@@ -67,6 +67,15 @@ test('skip-test', async (t) => {
     t.equal(testsuites.$.errors, '0')
     t.equal(testsuites.$.failures, '0')
     t.equal(testsuites.$.skipped, '1')
+    t.true(
+        testsuites.testsuite[0].testcase[0].skipped,
+        'first test marked as skipped'
+    )
+    t.false(
+        testsuites.testsuite[0].testcase[1].skipped,
+        'second test not marked as skipped'
+    )
+})
 })
 
 test('failure-in-beforeAll', async (t) => {